Rm multiply+divide to reduce search space.
Search time reduces to 1 second from 130 seconds on desktop on 3*3 grid for unsuccessful compilation. However, a multiply might actually be required. So this isn't a strict improvement.

On a 5*5 grid with successful compilation, compile time is ~4 seconds. On a 4*4 grid, I gave up after 108 minutes. Unsure why the result is not monotonic.