Bug fixes for synthesized and canonical allocations.  (#10)
* Due to .j2 file, we do not need the max here

in .j2 file, we will at most modify state_group_num

stateful vars per stage. But we have assert statement

to guarantee the sum of salu_config per stage will

be less than # of phvs per stage

* Remove unnecessary min

* Fix a bug in range() should only be self.num_state_groups

* No need to use salu_config for canonicalized allocation

* Modify the router_data_path_sketch.j2 file correspondingly

* Typo in comment

Thanks to Taegyun's good catch

* Rename the function so that we know which is for synthesized and which is for canonicalized

* Use salu_config for both canonicalized and synthesized

To be specific, in canonicalized allocation,

salu_config will be 2 dimension with assert constrain

in synthesized allocation, salu_config will be 3 dimension

because each decide which state group be allocated to which

stage and which alu in that stage

* Fix a typo in synthesized allocation credit to Anirudh's good catch

In synthesized allocation, each stateful_alu has one salu_config,

therefore, we make salu_config stick to the No. of stateful_alu

* Modify the initialization part of stateful_var in synthesized alloc

* Miss one endfor in .j2 file

* Make some changes to code generator for synthesized alloc