Fix bug in how tofino.alu's output alu is handled.
The salu ppt says that the output is either the updated value of
the register (alu_lo, alu_hi) or the original value (original_lo,
original_hi).

Note that original_lo and original_hi are called register_lo and
register_hi in the P4 code, but we use this for the updated value
of the registers because we need to update state in place in the .alu
files. Will fix this.