Several more bug fixes to Tofino code generation (#30)
* Rename file into something more appropriate.

* XORs aren't accepted by the Tofino compiler. Switch to AND for now.

* Fix bug in how tofino.alu's output alu is handled.

The salu ppt says that the output is either the updated value of
the register (alu_lo, alu_hi) or the original value (original_lo,
original_hi).

Note that original_lo and original_hi are called register_lo and
register_hi in the P4 code, but we use this for the updated value
of the registers because we need to update state in place in the .alu
files. Will fix this.

* Bug fix: needs to be signed; default in P4-14 is unsigned.

Unsigned leads to unexpected behavior:

https://wesmckinney.com/blog/avoid-unsigned-integers/
https://google.github.io/styleguide/cppguide.html#Integer_Types

* Add helper script to automate building and running for Tofino.

* State variables should be signed as well.

* Fix failing test.

* Ask user for test inputs to run.sh

* Simplify script; run should be decoupled from compile.

* Use md5hash to avoid long table names if sketch name is long.

bfshell doesn't seem to handle long table names well and seg
faults when executing control plane commands. We need to dig deeper into
this, but for now shortening the table names seems like a reasonable
workaround.

* The sketch_name should start with char rather than number

* Add fix for 8b1091f1deab85f3522476714c754f3c237b76ac

* Format files.

* Add comment about XOR operations.

* Add test for last commit.

* Add one more test case

* Add one more space in test_tofino file

Co-authored-by: Taegyun Kim <k.taegyun@gmail.com>
Co-authored-by: XiangyuG <39160477+XiangyuG@users.noreply.github.com>